pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zephyr-desktop"
version = "0.1.0"
description = "Desktop app for managing Ralph loops via Docker containers"
requires-python = ">=3.12"
license = {text = "MIT"}

dependencies = [
    "PyQt6>=6.6.0",
    "docker>=7.0.0",
    "gitpython>=3.1.40",
    "keyring>=25.0.0",
    "playwright>=1.40.0",
    "plyer>=2.1.0",
    "anthropic>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-qt>=4.3.0",
    "pytest-mock>=3.12.0",
    "black>=24.0.0",
    "pylint>=3.0.0",
    "pyyaml>=6.0",
    "pyinstaller>=6.0.0",
]

[project.scripts]
zephyr = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src", "src.*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
qt_api = "pyqt6"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.pylint.main]
source-roots = ["src"]
